requirements.txt
langchain
langchain-groq
langchain-community
faiss-cpu
sentence-transformers
PyPDF2
SpeechRecognition
gTTS
gradio
python-dotenv
